Update virtual zarr store examples (#36)
@emmanuelmathot I split the example into an item and a collection
example and made one of them a kerchunk and one a virtual icechunk
store.

These are based off of
https://docs.openveda.cloud/user-guide/notebooks/veda-operations/publish-cmip6-virtual-zarr-stac.html
which represents the current best thinking of how to represent these
types of data in STAC.

As I was writing this I was particularly unhappy with the pattern of
using the Virtual Asset STAC Extension to create the relationship
between the virtual Zarr store and the source data files. It might be
more explicit to make a new field in the Zarr STAC Extension to make
that relationship explicit.

Co-authored-by: Emmanuel Mathot <emmanuel.mathot@gmail.com>